backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
alembic==1.15.1
asyncpg==0.30.0
redis==5.2.1
celery==5.4.0
pydantic-settings==2.8.1
email-validator==2.2.0
python-multipart==0.0.20
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
aiosqlite==0.21.0
ruff==0.11.0
openpyxl==3.1.5
reportlab==4.2.5
locust==2.43.4
